seet/requirements.txt
transformers
tensorflow
django
djangorestframework
numpy
librosa